api types [nfc]: Move UserSettings to modelTypes.js
